Converting all floats to doubles

+++ src/python_bindings/fit_python.cpp
@@ -36,7 +36,7 @@ void km::KMedoidsWrapper::fitPython(
   if ((kw.size() != 0) && (kw.contains("k"))) {
     KMedoids::setNMedoids(pybind11::cast<int>(kw["k"]));
   }
-  KMedoids::fit(carma::arr_to_mat<float>(inputData), loss);
+  KMedoids::fit(carma::arr_to_mat<double>(inputData), loss);
 }
 
 void fit_python(pybind11::class_<KMedoidsWrapper> *cls) {
